make completeness changes trigger when save completes
having both set() and save() happen can mean that todos flicker between
being "done" and "not done" as set() triggers and event then a while
later save() triggers another one. Just tick the same todo on and off
a few times.

Note the behaviour might be faster in production but likely we'll have
to fix this anyway.